+++ lib/AppInfo/Application.php
@@ -5,11 +5,22 @@ declare(strict_types=1);
 namespace OCA\EnhancedRegistration\AppInfo;
 
 use OCP\AppFramework\App;
+use OCP\AppFramework\Bootstrap\IBootContext;
+use OCP\AppFramework\Bootstrap\IBootstrap;
+use OCP\AppFramework\Bootstrap\IRegistrationContext;
+use OCP\Util;
 
-class Application extends App {
+class Application extends App implements IBootstrap {
     public const APP_ID = "enhanced_registration";
 
     public function __construct() {
         parent::__construct(self::APP_ID);
+    }
+
+    public function register(IRegistrationContext $context): void {
+    }
+
+    public function boot(IBootContext $context): void {
+        Util::addScript(self::APP_ID, 'loginlink');
     }
 }